[wip] extract prismic preview code from common app
