+++ catalogue/webapp/pages/_app.tsx
@@ -1,5 +1,5 @@
-import NextApp, { NextWebVitalsMetric, AppContext } from 'next/app';
-import App, { WecoAppProps } from '@weco/common/views/pages/_app';
+import NextApp, { NextWebVitalsMetric, AppContext, AppProps } from 'next/app';
+import App from '@weco/common/views/pages/_app';
 import { SearchContextProvider } from '@weco/common/views/components/SearchContext/SearchContext';
 import { gtagReportWebVitals } from '@weco/common/utils/gtag';
 import { ReactElement } from 'react';
@@ -8,7 +8,7 @@ export function reportWebVitals(metric: NextWebVitalsMetric): void {
   gtagReportWebVitals(metric);
 }
 
-export default function CatalogueApp(props: WecoAppProps): ReactElement {
+export default function CatalogueApp(props: AppProps): ReactElement {
   return (
     <SearchContextProvider>
       <App {...props} />
