+++ common/services/app/analytics.ts
@@ -1,0 +1,3 @@
+import ReactGA from 'react-ga';
+
+const engagementTracker = () => {};
